requirements.txt
pytest==8.0.0
pytest-html==4.1.1
requests==2.31.0
python-dotenv==1.0.0
msal==1.26.0
python-graphql-client==0.4.3 